+++ AltecSystem.Application/DTOs/Auth/AddPointsRequestDto.cs
@@ -3,5 +3,5 @@ namespace AltecSystem.Application.DTOs.Auth;
 public class AddPointsRequestDto
 {
     public string Name { get; set; } = default!;
-    public int Points { get; set; }
+    public decimal Points { get; set; } 
 }
